[ADD]: Inventory

+++ src/main/java/com/vastworld/vwbe/dto/player/GetPlayerNextBreakthroughResponse.java
@@ -3,5 +3,5 @@ package com.vastworld.vwbe.dto.player;
 public record GetPlayerNextBreakthroughResponse(
     Boolean isTribulation,
     Long breakthroughPoints,
-    Float chanceOfSuccess
+    Double chanceOfSuccess
 ) {}
